pyproject.toml
[project]
name = "sample-django-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=4.2.23",
    "django-compressor>=4.5.1",
    "django-crispy-forms>=2.4",
    "django-environ>=0.11.2",
    "django-extensions>=3.2",
    "django-sass-processor>=1.4.2",
    "django-ses>=4.4.0",
    "django-storages>=1.14.6",
    "gunicorn>=23.0.0",
    "ipython>=8.12.3",
    "libsass>=0.23.0",
    "openai>=1.98.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "sentry-sdk>=2.34.1",
    "social-auth-app-django>=5.4.2",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.0",
    "coverage>=7.6.1",
    "django-debug-toolbar>=4.4.6",
    "flake8>=5.0.4",
    "ipdb>=0.13.13",
    "ipython>=8.12.3",
    "isort>=5.13.2",
    "pip-tools>=7.4.1",
    "pre-commit>=3.5.0",
    "rich>=14.1.0",
    "werkzeug>=3.0.6",
]
